ajout de la page detaile en stack avec etude + initialisation du routing avec routeur.push + ajout dossier (auth) a configurer)
