+++ app/(tabs)/etudes/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "expo-router";
+
+export default function EtudesLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false, // ⛔️ aucun header sur toutes les pages de ce dossier
+      }}
+    />
+  );
+}
