+++ package.json
@@ -39,7 +39,8 @@
     "react-native-screens": "~4.10.0",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5",
-    "expo-linear-gradient": "~14.1.4"
+    "expo-linear-gradient": "~14.1.4",
+    "expo-secure-store": "~14.2.3"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
